Call qmake from the root of the ${WRKSRC}.
In Qt 5.3, Qt's build infrastructure has undergone some changes that
make our previous approach of calling qmake from the directories we want
to build stop working. Things would break even more in Qt 5.4, in which
qtbase's configure script does not accept the -process, -fully-process
and -dont-process arguments anymore (it always behaves as if -process
had been used).

Bite the bullet and start calling qmake from ${WRKSRC}. The largest part
of this change involves changing lines in Makefiles from
  WRKSRC_SUBDIR=	foo/bar
to
  BUILD_WRKSRC=		${WRKSRC}/foo/bar
  INSTALL_WRKSRC=	${WRKSRC}/foo/bar
as well as adding patches to .pro files to avoid entering other
subdirectories and removing post-configure targets that are not
necessary anymore.

Since qmake needs to be called from the top of ${WRKSRC} anyway, we
can also simplify the configuration process for the qtbase ports a
little. Looking at r10019 it is not clear why we started calling
qmake in the pre-configure target in addition to the post-configure
one (while also skipping it in do-configure), but we can now drop
this call since letting configure behave as if -process had been
passed means it will call qmake on its own and overwrite the files
generated by the pre-configure call. We still need to call qmake in
post-configure though, as the configure script does not pass
-recursive when calling qmake and we need to be able to call make
from any subdirectory when building.