bsd.qt.mk: Set QMAKESPEC instead of QMAKEPATH.
This partially reverts r8728.

QMAKEPATH does much more than we want now that we call qmake from the
top of ${WRKSRC}. qmake uses QMAKEPATH when evaluating the QMAKE_MKSPECS
property, which is in turn used by qt_config.pri to load the .pri files
in mkspecs/modules.

In practice, this means that if people have an older Qt installation
those files will be used and QT_CONFIG will have values such as "gui"
even if one is building a port like textproc/qt5-xml, which passes
-no-gui to the configure script. Consequently, unintended code paths
may be enabled or the configuration step can just fail if the .pro files
expect values that are not present in the system-wide, older .pri files.

We avoid all those problems if we use QMAKESPEC, as qmake does not take
its value into account when evaluating the QMAKE_MKSPECS property and
will only parse the files in the mkspec's
directory (mkspecs/freebsd-clang, for example, instead of all the files
in mkspecs).