Merge devel/qt5-qmldevtools into lang/qt5-qml.
qt5-qmldev only worked as a standalone port with a lot of duct tape and
our hacks that invoked qmake from a subdirectory of ${WRKSRC}.

Since the latter does not work anymore, we cannot build qt5-qmldevtools
as we did before: the "no_module_headers" sed call in the Makefile no
longer works, qt5-qml is needed as a build-time dependency and the
headers (which actually belong to qt5-qml and are also installed by it
into a different location) cannot be installed anymore.

Since the port now depends on qt5-qml to build and the header files it
used to install are now shipped only by qt5-qml, there is no purpose in
keeping the port around.

ack'ed by makc@ and avilla@.