qt5-core: Remove -no-feature-* entries which do not exist.
Starting with Qt 5.4, the configure script rejects -no-feature-* arguments
for features which are not listed in src/corelib/global/qfeatures.txt. This
is the case for:

 * -no-feature-alsa
 * -no-feature-evdev
 * -no-feature-glib
 * -no-feature-pulseaudio
 * -no-feature-inotify
 * -no-feature-libudev
 * -no-feature-xcb

Most of those arguments can be turned into "-no-<feature>", which in most
cases have pretty much the same effect.

-no-feature-glib has been removed altogether, as -no-glib means glib support
will not be enabled at all. -no-libudev and -no-xcb does not the QT_LIBUDEV
and QT_XCB blocks to qconfig.h, but this should hopefully not be a problem.