Update patches, distinfo and plists for Qt 5.4.1.
In an attempt to reduce the number of commits during which the tree is
broken, this commit does most of the required work to update Qt to a new
release and updates things such as plists, distinfo, patches and the
Makefiles (to stop processing directories that do not exist anymore etc).

This is actually most of the work to move to 5.4.1: the next commits will
just make some ports build in older FreeBSD releases and fixes qt5-gui.

Big thanks to Alex Richardson <aricharson.kde@gmail.com> and Tobias Berner
<tcberner@gmail.com> for help with testing and plist patches!