Start making qt5-gui depending on print/harfbuzz.
With 5.4.0, HarfBuzz-NG is Qt's default shaper, as described in QTBUG-18980
and in the release notes.

We can leverage this change and start using the version in ports instead of
the one bundled with Qt, as bundled dependencies should be avoided as much
as possible. Additionally, the build was failing with the bundled HarfBuzz
and we would need to add more hacks to the Makefile to build it before
src/gui.